Squash warning

+++ src/tls_openssl.c
@@ -143,7 +143,7 @@ void mg_tls_free(struct mg_connection *c) {
 
 size_t mg_tls_pending(struct mg_connection *c) {
   struct mg_tls *tls = (struct mg_tls *) c->tls;
-  return tls == NULL ? 0 : SSL_pending(tls->ssl);
+  return tls == NULL ? 0 : (size_t) SSL_pending(tls->ssl);
 }
 
 long mg_tls_recv(struct mg_connection *c, void *buf, size_t len) {
